+++ src/modules/canvas/hooks/index.ts
@@ -1,2 +1,1 @@
-// Canvas hooks - to be implemented in future WPs
-export {};
+export { useNotesAsNodes } from './useNotesAsNodes';